+++ server/controllers/auth/logout/index.js
