+++ server/controllers/auth/signup/index.js
