--- server/controllers/fallback/sendReactApp.js
@@ -1,7 +1,0 @@
-const handlePageRender = require('../../render/build/handlePageRender.js');
-
-const sendReactApp = (req, res) => {
-  handlePageRender(req, res);
-};
-
-module.exports = sendReactApp;
