+++ server/routes/fallback/index.js
@@ -1,4 +1,4 @@
-const handlePageRequest = require('../../controllers/fallback/sendReactApp');
+const handlePageRequest = require('../../controllers/pages/sendReactApp');
 
 module.exports = (app) => {
   app.get('*', handlePageRequest);
